Fixed hanging end-marker in AddPair function. Now, when a begin is added, it is also added to the end if no end is supplied.
If an end is later supplied, it overwrites the setting. To make sure garbage collection is done correctly, Deallocate now
checks the end marker for equality to the beginning marker. If not the same, deallocate both, otherwise, just deallocate
the begin marker. The begin marker is deallocated last to ensure no indeterminate state of the pointer.